Break out styles, setup handle functions
